refactor(web): reorganize src into domain-driven folders
Move web/src from horizontal technical-type grouping (components/,
hooks/, lib/) into vertical, domain-driven folders so each domain's
components, hooks, and helpers sit under one domain-named folder.

- chat/ (with sort/), conversation/, metadata/, agent/, shared/ (with ui/)
- rename chatSort.ts -> sortConfig.ts (exported names unchanged)
- repoint components.json aliases to shared/ (ui -> @/shared/ui)
- all moves via git mv; test files travel with their subjects

Pure move + import-fix + one rename: no runtime behavior or component
logic changes. typecheck, 107 unit tests, build, and e2e all pass.

Closes #103
