+++ web/src/chat/FilterPanel.tsx
