+++ web/src/conversation/CollapsibleThinking.tsx
