+++ web/src/shared/ui/button.tsx
@@ -1,7 +1,7 @@
 import { Button as ButtonPrimitive } from "@base-ui/react/button";
 import { type VariantProps } from "class-variance-authority";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/utils";
 import { buttonVariants } from "./button-variants";
 
 function Button({
