+++ web/src/shared/ui/resizable.tsx
@@ -1,6 +1,6 @@
 import * as ResizablePrimitive from "react-resizable-panels";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/utils";
 
 function ResizablePanelGroup({
   className,
